requirements.txt
streamlit
pandas
feedparser
requests
beautifulsoup4
mysql-connector-python
openai
